setu-aa-server/requirements.txt
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==35.0.0
dnspython==2.1.0
Flask==2.0.2
google-api-core==2.1.0
google-auth==2.3.0
google-cloud-pubsub==2.8.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pendulum==2.1.2
proto-plus==1.19.3
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytzdata==2020.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
